fix: Remove unnecessary i64 casts in helium-server
- Remove unnecessary casts for gas_wanted and gas_used in abci_server.rs
- Remove unnecessary casts for gas_wanted and gas_used in services/tx.rs
- These fields are already i64 in the proto-generated TxResponse type

+++ crates/helium-server/src/abci_server.rs
@@ -375,8 +375,8 @@ impl AbciService for AbciServer {
             data: vec![],
             log: result.log,
             info: String::new(),
-            gas_wanted: result.gas_wanted as i64,
-            gas_used: result.gas_used as i64,
+            gas_wanted: result.gas_wanted,
+            gas_used: result.gas_used,
             events: convert_events(result.events),
             codespace: String::new(),
         }))
